git-svn-id: https://10.8.9.8/svn/Project_ARMLinux/trunk/kernel/v2.6.34@460 34fca4b9-5ec2-4a8b-9ba4-7e5d7900cddd

+++ drivers/amlogic/tvin/bt656_601_in.c
@@ -55,8 +55,8 @@
 
 //static struct am656in_dev_t am656in_dev_;
 typedef struct {
-    unsigned long   pbufAddr;
-    unsigned long   pbufSize;
+    unsigned        pbufAddr;
+    unsigned        pbufSize;
 
     unsigned        pin_mux_reg1;   //for bt656 input
     unsigned        pin_mux_mask1;
